+++ packages/error/src/UnimplementedError.ts
@@ -1,8 +1,7 @@
-import { Ambiguous } from '@jamashita/anden-type';
 import { RuntimeError } from './RuntimeError';
 
 export class UnimplementedError extends RuntimeError {
-  public constructor(message: string = 'UNIMPLEMENTED', cause?: Ambiguous<Error>) {
+  public constructor(message: string = 'UNIMPLEMENTED', cause?: Error) {
     super(message, cause);
   }
 }