+++ packages/zeit/src/ZeitError.ts
@@ -1,8 +1,7 @@
 import { RuntimeError } from '@jamashita/anden-error';
-import { Ambiguous } from '@jamashita/anden-type';
 
 export class ZeitError extends RuntimeError {
-  public constructor(message: string, cause?: Ambiguous<Error>) {
+  public constructor(message: string, cause?: Error) {
     super(message, cause);
   }
 }